EWD-23742 - Deleted learning content should not try to be saved
Former-commit-id: a6a7bf6da97ba4124f2fc7523f98d5f58a4aec83

+++ sources/easygenerator.Web/App/viewmodels/questions/learningContents.js
@@ -54,6 +54,11 @@
         function removeLearningContent(learningContent) {
             eventTracker.publish(events.deleteLearningContent);
 
+            if (!_.isNullOrUndefined(learningContent.isDeleted) && learningContent.isDeleted) {
+                viewModel.learningContents.remove(learningContent);
+                return;
+            }
+
             performActionWhenLearningContentIdIsSet(learningContent, function () {
                 viewModel.learningContents.remove(learningContent);
                 repository.removeLearningContent(viewModel.questionId, ko.unwrap(learningContent.id)).then(function (response) {
@@ -91,7 +96,7 @@
             var id = ko.unwrap(learningContent.id);
             var text = ko.unwrap(learningContent.text);
 
-            if (_.isEmptyHtmlText(text)) {
+            if (_.isEmptyHtmlText(text) || ((!_.isNullOrUndefined(learningContent.isDeleted) && learningContent.isDeleted))) {
                 return;
             }
